Mostrando erro ao tentar deletar autor com livro relacionado, em comentário no código há a resolução para excluir o livro relacionado

+++ src/br/com/caelum/livraria/bean/AutorBean.java
@@ -3,11 +3,15 @@ package br.com.caelum.livraria.bean;
 import java.io.Serializable;
 import java.util.List;
 
+import javax.faces.application.FacesMessage;
+import javax.faces.context.FacesContext;
 import javax.inject.Inject;
 import javax.inject.Named;
 
 import br.com.caelum.livraria.dao.AutorDao;
+import br.com.caelum.livraria.dao.LivroDao;
 import br.com.caelum.livraria.modelo.Autor;
+import br.com.caelum.livraria.modelo.Livro;
 import br.com.caelum.livraria.tx.Transactional;
 
 @Named
@@ -15,13 +19,15 @@ import br.com.caelum.livraria.tx.Transactional;
 public class AutorBean implements Serializable {
 
 	private static final long serialVersionUID = 1L;
-	
+
 	@Inject
 	private AutorDao dao;
+	@Inject
+	LivroDao livroDao;
 	private Autor autor = new Autor();
-	
+
 	private Integer autorId;
-	
+
 	public AutorBean() {
 	}
 
@@ -32,7 +38,7 @@ public class AutorBean implements Serializable {
 	public void setAutorId(Integer autorId) {
 		this.autorId = autorId;
 	}
-	
+
 	public void carregarAutorPelaId() {
 		this.autor = this.dao.buscaPorId(autorId);
 	}
@@ -41,7 +47,7 @@ public class AutorBean implements Serializable {
 	public String gravar() {
 		System.out.println("Gravando autor " + this.autor.getNome());
 
-		if(this.autor.getId() == null) {
+		if (this.autor.getId() == null) {
 			this.dao.adiciona(this.autor);
 		} else {
 			this.dao.atualiza(this.autor);
@@ -54,14 +60,27 @@ public class AutorBean implements Serializable {
 
 	@Transactional
 	public void remover(Autor autor) {
+		List<Livro> livrosDoAutor = this.dao.listaTodosLivros(autor.getId());
+		if (!livrosDoAutor.isEmpty()) {
+			/*
+			 * for (Livro livro : livrosDoAutor) { livroDao.remove(livro); }
+			 */
+
+			for (Livro livro : livrosDoAutor)
+				System.out.println(livro.getTitulo());
+			FacesContext.getCurrentInstance().addMessage("autor",
+					new FacesMessage("Primeiro remova os livros do autor: ".concat(livrosDoAutor.toString())));
+			return;
+
+		}
 		System.out.println("Removendo autor " + autor.getNome());
 		this.dao.remove(autor);
 	}
-	
+
 	public List<Autor> getAutores() {
 		return this.dao.listaTodos();
 	}
-	
+
 	public Autor getAutor() {
 		return autor;
 	}
